end server template(without InfoWriterHandler logic))

+++ InfoWriterWebSocketServer/InfoWriterWebSocketServer/Models/Update.cs
@@ -10,7 +10,7 @@ namespace InfoWriterWebSocketServer.Models
     {
         public FrameMessageEnum  Frame { get; set; }
         public string Payload { get; set; }
-        public int PayloadLength { get; set; }
+        public long PayloadLength { get; set; }
         public byte[] Mask { get; set; } 
         public bool IsFullUpdate { get; set; }
     }
